+++ src/components/common/StudentCourseRegister/StudentRegistration.tsx
@@ -1,5 +1,6 @@
+import StudentRegisterContent from './RegiserContent'
 import StudentRegisterModal from './RegiserModal'
-import StudentRegisterContent from '../../Login/RegiserContent'
+
 import { useState } from 'react'
 
 // //수강생 등록 모달 내용
